f20 - hopefully fixes issue with dms and long outputs. Refactored some copy/pasted methods into a base interaction module class.
