+++ Gao.Gurps.Discord/Slash/DiceRollingModule.cs
@@ -22,49 +22,43 @@ namespace Gao.Gurps.Discord.Slash
 #if DEBUG
     [Group("debug-dice", "test commands")]
 #endif
-    public class DiceRollingModule : InteractionModuleBase
+    public class DiceRollingModule : GurpsInteractionModuleBase
     {
-
-        private async Task SendTextWithTimeBuffers(int linesToDisplayAtOnce, IEnumerable<string> results, bool privateMessageOverflow, string lineDivider = "")
+                [SlashCommand("gather-energy", "Gather energy via the Ritual Path Magic mechanics", runMode: RunMode.Async)]
+        [CommandSummary(@"Gather Energy according to the rules of RPM. 
+Example command usage: 
+To gather 30 energy with 15 skill.
+\gather-energy Path of Mind-15 30
+To gather 30 energy with 15 skill and gathering occurs every 5 seconds.
+\gather-energy Path of Body-15 30 00:00:05
+If you are in a hurry to gather 30 energy with 15 skill.
+\gather-energy 15 30 ")]
+        public async Task GatherEnergy(string argument)
         {
-            if (lineDivider == string.Empty) lineDivider = Environment.NewLine;
-            const int maxMessageSize = 2000 - 30; //Let's give some buffer room.
-            results = results.SelectMany(r => r.Length > maxMessageSize ? r.Split(lineDivider) : new[] { r }).Where(s => !string.IsNullOrWhiteSpace(s)).ToArray();
-            IMessageChannel overflowChannel;
-            var thisChannel = await Context.Guild.GetTextChannelAsync(Context.Interaction.ChannelId.Value);
-            if (privateMessageOverflow) overflowChannel = await Context.User.CreateDMChannelAsync();
-            else overflowChannel = thisChannel;
-            var remainingResults = results.ToArray();
-            var overflow = false;
-            IMessageChannel channel = thisChannel;
+            const int linesToDisplayAtOnce = 5;
 
-            while (remainingResults.Length > 0)
+            EnergyGatheringParameters parsedParameters;
+            try
             {
-                channel = overflow ? overflowChannel : thisChannel;
-                var sb = new StringBuilder();
-                sb.AppendLine("```");
-                var linesLeftToAppend = Math.Min(linesToDisplayAtOnce, remainingResults.Length);
-                var linesTaken = 0;
-                while (sb.Length + (remainingResults.FirstOrDefault() ?? "").Length < maxMessageSize && linesLeftToAppend > 0)
-                {
-                    sb.Append(remainingResults.First() + lineDivider);
-                    remainingResults = remainingResults.Skip(1).ToArray();
-                    linesLeftToAppend--;
-                    linesTaken++;
-                }
-                var displayValue = sb.ToString();
-                displayValue = displayValue.Trim(lineDivider.ToCharArray()) + "```";
-
-                displayValue = displayValue.Replace("``````", string.Empty);
-                await channel.SendMessageAsync(displayValue);
-                if (remainingResults.Length == 0)
-                    break;
-
-                await Task.Delay(1500);
-
-                overflow = true;
+                parsedParameters = EnergyGather.Parse(argument);
             }
+            catch (Exception ex)
+            {
+                await DisplayErrorMessage(await Context.Guild.GetTextChannelAsync(Context.Interaction.ChannelId.Value), ex.Message);
+                return;
+            }
+
+            var results = EnergyGather.Execute(parsedParameters);
+            if (parsedParameters.Verbose)
+            {
+                await Context.Interaction.RespondAsync("Starting to gather energy.");
+                await SendTextWithTimeBuffers(linesToDisplayAtOnce, results.VerboseOutput().Split(Environment.NewLine), false);
+            }
+            else
+                await Context.Interaction.RespondAsync(results.ToString());
+
         }
+
 
         [SlashCommand("quick-contest", "Execute a quick contest", runMode: RunMode.Async), CommandSummary(@"Roll Quick Contest.
 Examples:
